# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Every non-2xx response in this API has this shape.
 */
export interface ApiError {
  /** Stable machine-readable code, e.g. `table_taken`. */
  error: string;
  /** Human-readable Thai text, safe to show to a guest. */
  message: string;
}

/**
 * Seating slot, HH:mm.
 */
export type Slot = typeof Slot[keyof typeof Slot];


export const Slot = {
  '18:00': '18:00',
  '19:00': '19:00',
  '20:00': '20:00',
  '21:00': '21:00',
  '22:00': '22:00',
  '23:00': '23:00',
} as const;

export type TableShape = typeof TableShape[keyof typeof TableShape];


export const TableShape = {
  round: 'round',
  rect: 'rect',
} as const;

export type OrderKind = typeof OrderKind[keyof typeof OrderKind];


export const OrderKind = {
  table: 'table',
  ticket: 'ticket',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
  expired: 'expired',
} as const;

export interface Zone {
  id: string;
  name: string;
  description: string;
  /** Minimum spend in THB. 0 means none. */
  minSpend: number;
  capacityMin: number;
  capacityMax: number;
}

export interface FloorTable {
  id: string;
  zoneId: string;
  shape: TableShape;
  x: number;
  y: number;
  w: number;
  h: number;
  minSeats: number;
  maxSeats: number;
}

export interface FloorPlan {
  zones: Zone[];
  tables: FloorTable[];
}

export type VenueEventKind = typeof VenueEventKind[keyof typeof VenueEventKind];


export const VenueEventKind = {
  concert: 'concert',
  dj: 'dj',
  festival: 'festival',
  special: 'special',
} as const;

export interface VenueEvent {
  id: string;
  kind: VenueEventKind;
  title: string;
  artistTh: string;
  artistEn: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  doorsAt: string;
  /** Price per ticket in THB. */
  price: number;
  capacity: number;
  ticketsLeft: number;
}

export interface TableAvailability {
  id: string;
  free: boolean;
  /** Free *and* the party fits, when `guests` was supplied. */
  selectable: boolean;
}

export interface Availability {
  date: string;
  slot: Slot;
  freeTables: number;
  tables: TableAvailability[];
}

/**
 * One shape for both kinds. `guests`/`tableId`/`zoneId`/`zoneName` are present only for table bookings; `eventId`/`eventTitle`/`quantity` only for ticket orders.
 */
export interface Order {
  code: string;
  kind: OrderKind;
  status: OrderStatus;
  bookerName: string;
  phone: string;
  email?: string;
  occasion?: string;
  notes?: string;
  date: string;
  slot: string;
  holdUntil: string;
  guests?: number;
  tableId?: string;
  zoneId?: string;
  zoneName?: string;
  eventId?: string;
  eventTitle?: string;
  quantity?: number;
  /** What the guest pays, VAT included, in THB. */
  amount: number;
  /** VAT-exclusive portion of `amount`. */
  amountBase: number;
  amountVat: number;
  holdExpiresAt: string;
  createdAt: string;
  paidAt?: string;
  cancelledAt?: string;
}

export interface CreateBookingRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  bookerName: string;
  /**
     * @minLength 9
     * @maxLength 20
     */
  phone: string;
  email?: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  guests: number;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  slot?: Slot;
  tableId: string;
  /** @pattern ^\d{2}:\d{2}$ */
  holdUntil?: string;
  /** @maxLength 60 */
  occasion?: string;
  /** @maxLength 500 */
  notes?: string;
}

export interface CreateTicketOrderRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  bookerName: string;
  /**
     * @minLength 9
     * @maxLength 20
     */
  phone: string;
  email?: string;
  eventId: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  quantity: number;
  /** @maxLength 500 */
  notes?: string;
}

export interface SubmitSlipRequest {
  /**
     * Lowercase hex SHA-256 of the slip image.
     * @pattern ^[0-9a-f]{64}$
     */
  hash: string;
  /** @maxLength 255 */
  fileName: string;
  /**
     * @minimum 1
     * @maximum 5242880
     */
  sizeBytes: number;
}

export type SlipResultSlipStatus = typeof SlipResultSlipStatus[keyof typeof SlipResultSlipStatus];


export const SlipResultSlipStatus = {
  pending_bank_check: 'pending_bank_check',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface SlipResult {
  order: Order;
  slipStatus: SlipResultSlipStatus;
  /** True until a slip-verification provider is wired up. The duplicate check is real; authenticity is not yet checked. */
  pendingBankCheck: boolean;
}

export interface CancelOrderRequest {
  /**
     * @minLength 9
     * @maxLength 20
     */
  phone: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  member: 'member',
  admin: 'admin',
} as const;

/**
 * Never carries the password hash or the session token.
 */
export interface AuthUser {
  id: string;
  name: string;
  phone: string;
  email: string;
  role: UserRole;
  suspended: boolean;
  createdAt: string;
  lastLoginAt?: string;
}

export interface RegisterRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 9
     * @maxLength 20
     */
  phone: string;
  /**
     * @minLength 5
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]{2,}$
     */
  email: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  password: string;
  /** Must be true. The server records when it was given. */
  pdpaConsent: boolean;
}

/**
 * Only the fields present are changed.
 */
export interface UpdateProfileRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name?: string;
  /**
     * @minLength 5
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]{2,}$
     */
  email?: string;
}

export interface ChangePasswordRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  newPassword: string;
}

export interface LoginRequest {
  /**
     * @minLength 9
     * @maxLength 20
     */
  phone: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  password: string;
}

/**
 * Not signed in, or the credentials were wrong
 */
export type UnauthorizedResponse = ApiError;

/**
 * The request was rejected
 */
export type BadRequestResponse = ApiError;

/**
 * No such resource
 */
export type NotFoundResponse = ApiError;

/**
 * The request conflicts with the current state
 */
export type ConflictResponse = ApiError;

export type GetAvailabilityParams = {
/**
 * Service date, yyyy-MM-dd
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
date: string;
/**
 * Seating slot, HH:mm. Defaults to the venue's main seating.
 */
slot?: Slot;
/**
 * Party size, used to compute `selectable`.
 * @minimum 1
 * @maximum 20
 */
guests?: number;
};

export type ListOrdersParams = {
/**
 * Thai mobile number; punctuation is ignored.
 * @minLength 9
 * @maxLength 20
 */
phone: string;
kind?: OrderKind;
};

export type ListMyOrdersParams = {
kind?: OrderKind;
};