# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      description: >
        Creates the account and signs it in — the response sets an httpOnly
        session cookie. PDPA consent is required, and the time it was given is
        recorded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: The signed-in account
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Change your own name or email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change your own password
      description: >
        Every other session is revoked, so a password changed because it leaked
        actually locks the other party out.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Changed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/orders:
    get:
      operationId: listMyOrders
      tags: [auth]
      summary: Your own bookings and tickets
      description: >
        Backs "My tickets" for a signed-in member. Matches on the account's
        phone number, so bookings made before signing up still appear.
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderKind"
      responses:
        "200":
          description: Orders, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  responses:
    Unauthorized:
      description: Not signed in, or the credentials were wrong
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    BadRequest:
      description: The request was rejected
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    Conflict:
      description: The request conflicts with the current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      description: Every non-2xx response in this API has this shape.
      properties:
        error:
          type: string
          description: Stable machine-readable code, e.g. `table_taken`.
        message:
          type: string
          description: Human-readable Thai text, safe to show to a guest.
      required:
        - error
        - message
...
    OrderKind:
      type: string
      enum: [table, ticket]
    OrderStatus:
      type: string
      enum: [pending, paid, cancelled, expired]
...
    Order:
      type: object
      description: >
        One shape for both kinds. `guests`/`tableId`/`zoneId`/`zoneName` are
        present only for table bookings; `eventId`/`eventTitle`/`quantity` only
        for ticket orders.
      properties:
        code:
          type: string
        kind:
          $ref: "#/components/schemas/OrderKind"
        status:
          $ref: "#/components/schemas/OrderStatus"
        bookerName:
          type: string
        phone:
          type: string
        email:
          type: string
        occasion:
          type: string
        notes:
          type: string
        date:
          type: string
        slot:
          type: string
        holdUntil:
          type: string
        guests:
          type: integer
        tableId:
          type: string
        zoneId:
          type: string
        zoneName:
          type: string
        eventId:
          type: string
        eventTitle:
          type: string
        quantity:
          type: integer
        amount:
          type: number
          description: What the guest pays, VAT included, in THB.
        amountBase:
          type: number
          description: VAT-exclusive portion of `amount`.
        amountVat:
          type: number
        holdExpiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
      required:
        - code
        - kind
        - status
        - bookerName
        - phone
        - date
        - slot
        - holdUntil
        - amount
        - amountBase
        - amountVat
        - holdExpiresAt
        - createdAt
...
    UserRole:
      type: string
      enum: [member, admin]
    AuthUser:
      type: object
      description: Never carries the password hash or the session token.
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        suspended:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - email
        - role
        - suspended
        - createdAt
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 9
          maxLength: 20
        email:
          # A pattern, not `format: email`: Orval renders that as zod.email(),
          # which is zod v4 syntax and this workspace is on v3.
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]{2,}$"
          minLength: 5
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 200
        pdpaConsent:
          type: boolean
          description: Must be true. The server records when it was given.
      required:
        - name
        - phone
        - email
        - password
        - pdpaConsent
    UpdateProfileRequest:
      type: object
      description: Only the fields present are changed.
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          # A pattern, not `format: email`: Orval renders that as zod.email(),
          # which is zod v4 syntax and this workspace is on v3.
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]{2,}$"
          minLength: 5
          maxLength: 254
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 200
        newPassword:
          type: string
          minLength: 8
          maxLength: 200
      required:
        - currentPassword
        - newPassword
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
          minLength: 9
          maxLength: 20
        password:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - phone
        - password