# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

/**
 * Never carries the password hash or the session token.
 */
export interface AuthUser {
  id: string;
  name: string;
  phone: string;
  email: string;
  role: UserRole;
  suspended: boolean;
  createdAt: Date;
  lastLoginAt?: Date;
}