# lib/api-zod/src/generated/types/changePasswordRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChangePasswordRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  newPassword: string;
}