# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 9
     * @maxLength 20
     */
  phone: string;
  /**
     * @minLength 5
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]{2,}$
     */
  email: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  password: string;
  /** Must be true. The server records when it was given. */
  pdpaConsent: boolean;
}