# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Only the fields present are changed.
 */
export interface UpdateProfileRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name?: string;
  /**
     * @minLength 5
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]{2,}$
     */
  email?: string;
}